requirements.txt
httpx==0.27.0
beautifulsoup4==4.13.3
python-dotenv==1.1.0
lxml==5.3.2
pandas==2.2.3
supabase==2.4.0
psycopg2-binary==2.9.9